+++ src/pandemia/components/movement_model/void_movement_model.py
@@ -22,16 +22,16 @@ class VoidMovementModel(MovementModel):
         number_of_agents = vector_region.number_of_agents
         number_of_activities = vector_region.number_of_activities
 
-        vector_region.requested_location_update = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.requesting_location_update = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.home_location = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.current_quarantine = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.current_location = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.current_facemask = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.requested_facemask_update = np.zeros((number_of_agents), dtype=np.int32)
-        vector_region.requesting_facemask_update = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.requested_location_update = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.requesting_location_update = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.home_location = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.current_quarantine = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.current_location = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.current_facemask = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.requested_facemask_update = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.requesting_facemask_update = np.zeros((number_of_agents), dtype=np.int64)
         vector_region.wears_facemask =\
-            np.zeros((number_of_agents, number_of_activities), dtype=np.int32)
+            np.zeros((number_of_agents, number_of_activities), dtype=np.int64)
 
     def initial_conditions(self, vector_region, offset):
